requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
PyYAML==6.0.2
pytest==8.3.4
httpx==0.28.1